chore(server): contine on with bugfixing (#865)
* chore(server): fix re-organising printers

* chore(server): fix and investigate further why that isn't working

* chore(server): sanitize id lists

* chore(server): clean up some elements and functions

* chore(server): change house keeping to 5 days not 1

* chore(server):fix bug

* chore(server): fix eval insecurity

* chore(server): how this not fix it... grr

* chore(client): upgrade to font-awesome 6

* chore(client): forgot these

* chore(client): i can copy and paste like a pro

* chore(client): fix re-writing global scope

* feat(client): file queue modal

* feat(client): file manager clean up and initial split

* feat(client): start abstraction of SSE client

* feat(client): file manager events connection and server sse service

* refactor(server): move amialive over to events endpoint

* refactor(main): remove those gash badges

* refactor(main): not needed

* feat(client): trigger client event for file data change

* refactor(main): commit work done

* Remove confusing readme stuff

* re-write initalise printers and create an enable printer queue

* file manager further clean up and prepwork

* add queue for enabling printers and offline check of printers

* increase offline check timers

* fix ui display for "searching"

* clean up and update packages

* tweakeroos to offline/online printer checking

* tasks tweask

* grab allowed reconnect states...

* fix printer onboarding

* fi enable/disable and swap to dropdown for manage functions

* fix quick connect state for disabled

* fix disabled button state

* fix enable/disable

* fix enabled state not persisting

* fix user lists

* split out modal for editing printer settings so can be changed with disabled

* remove unused import

* add printer edit modal, fix multi-users, allow alert to open edit

* fix plugin install/enable/disable list creation

* fix restart required never clearing

* clean up file-manager

* remove tags

* clean up some files

* remove some bugs

* fix printer onboarding again, and reconnecting to offline printers

* fix printer naming

* fix websocket reconnect

* fix websocket setting api online

* revert to passport 0.4.1 - 0.5.2 doesn't like my setup

* generate all file lists on demand

* stop file manager killing itself

* fix file reflow

* Revert "fix file reflow"

This reverts commit c8ecc4df6d5cb536e076910ba92e120408374e07.

* actual fix file reflow

* boom house keeping

* code of conduct - new readme layout - contributing samples to work with

* fix upload queue element trigger

* fix upload queue element trigger

* variable name

* delete all

* disable updating file information for now

* fix printer name not updating

* fix printer name not updating and generation

* file bugs begone!

* last bug in delete all... remove all folders

* weird css issue

* fix weird css issue

* fix scroll on houseclean

* fix notifications

* show no active printers message on file manager

* hide "Searching..." from views

* hide "Searching..." from views

* sprt disconnected button response to state

* fixes #871

* tweak current op view sizing better

* fix update check

* fix not goof on file manaer

* update AUTO message on action button quick connect

* live update file information

* don't always reload file folder

* missed a notification

* clean up logging

* starting history work

* starting history work

* cancelled and failed history

* disable some unfinished fields

+++ server/routes/settings.js
@@ -32,6 +32,7 @@ const S_VALID = require("../constants/validate-settings.constants");
 const { validateParamsMiddleware } = require("../middleware/validators");
 const M_VALID = require("../constants/validate-mongo.constants");
 const { sanitizeString } = require("../utils/sanitize-utils");
+const { databaseNamesList } = require("../constants/database.constants");
 
 module.exports = router;
 
@@ -70,7 +71,7 @@ router.delete(
   ensureAdministrator,
   async (req, res) => {
     try {
-      const deletedLogs = Logs.clearLogsOlderThan(1);
+      const deletedLogs = Logs.clearLogsOlderThan(5);
       res.send(deletedLogs);
     } catch (e) {
       logger.error("Failed to clear logs!", e);
@@ -103,7 +104,7 @@ router.post(
     // Will use in a future update to configure the dump.
     // let settings = req.body;
     // Generate the log package
-    let zipDumpResponse = {
+    const zipDumpResponse = {
       status: "error",
       msg: "Unable to generate zip file, please check 'OctoFarm-API.log' file for more information.",
       zipDumpPath: ""
@@ -150,14 +151,16 @@ router.get(
       await ProfilesDB.deleteMany({});
       logger.info("Successfully deleted Filament database.... Restarting server...");
       await SystemCommands.rebootOctoFarm();
-    } else {
+    } else if (databaseNamesList.includes(databaseName)) {
       await eval(databaseName).deleteMany({});
-      res.send({
-        message: "Successfully deleted " + databaseName + ", server will restart..."
-      });
-      logger.info(databaseName + " successfully deleted.... Restarting server...");
-      await SystemCommands.rebootOctoFarm();
+    } else {
+      logger.error("Unknown DB Name", databaseName);
     }
+    res.send({
+      message: `Successfully deleted ${databaseName}, server will restart...`
+    });
+    logger.info(databaseName + " successfully deleted.... Restarting server...");
+    await SystemCommands.rebootOctoFarm();
   }
 );
 router.get(
@@ -168,11 +171,11 @@ router.get(
   async (req, res) => {
     const databaseName = req.paramString("databaseName");
     logger.info("Client requests export of " + databaseName);
-    let returnedObjects = [];
+    const returnedObjects = [];
     if (databaseName === "FilamentDB") {
       returnedObjects.push(await ProfilesDB.find({}));
       returnedObjects.push(await SpoolsDB.find({}));
-    } else {
+    } else if (databaseNamesList.includes(databaseName)) {
       returnedObjects.push(await eval(databaseName).find({}));
     }
     logger.info("Returning to client database object: " + databaseName);
@@ -199,7 +202,7 @@ router.post(
       statusTypeForUser: "error",
       message: ""
     };
-    let force = req.bodyBool;
+    const force = req.body;
     if (
       !force ||
       typeof force?.forcePull !== "boolean" ||
@@ -237,7 +240,7 @@ router.post("/client/update", ensureCurrentUserAndGroup, ensureAuthenticated, as
   const currentUserList = await fetchUsers();
 
   // Patch to fill in user settings if it doesn't exist
-  for (let user of currentUserList) {
+  for (const user of currentUserList) {
     if (!user.clientSettings) {
       user.clientSettings = new ClientSettingsDB();
       user.clientSettings.save();
@@ -318,7 +321,7 @@ router.post("/server/update", ensureAuthenticated, ensureAdministrator, (req, re
     //Check the influx export to see if all information exists... disable if not...
     let shouldDisableInflux = false;
     let returnMsg = "";
-    let influx = sentOnline.influxExport;
+    const influx = sentOnline.influxExport;
     if (influx.active) {
       if (influx.host.length === 0) {
         shouldDisableInflux = true;
@@ -368,7 +371,7 @@ router.get(
   }
 );
 router.post("/customGcode/edit", ensureAuthenticated, async (req, res) => {
-  let script = await GcodeDB.findById(req.bodyString("id"));
+  const script = await GcodeDB.findById(req.bodyString("id"));
   script.gcode = req.bodyString("gcode");
   script.name = req.bodyString("name");
   script.description = req.bodyString("description");
@@ -385,7 +388,7 @@ router.post("/customGcode/edit", ensureAuthenticated, async (req, res) => {
 });
 
 router.post("/customGcode", ensureAuthenticated, async (req, res) => {
-  let newScript = req.body;
+  const newScript = req.body;
   const saveScript = new GcodeDB(newScript);
   saveScript
     .save()
@@ -403,7 +406,7 @@ router.get(
   async (req, res) => {
     const printerId = req.paramString("id");
     const all = await GcodeDB.find();
-    let returnCode = [];
+    const returnCode = [];
     all.forEach((script) => {
       if (script.printerIds.includes(printerId) || script.printerIds.length === 0) {
         returnCode.push(script);
